requirements.txt
boto3==1.4.3
pycrypto==2.6.1
